move freactal functions
