+++ packages/xo-web/src/xo-app/settings/remotes/remote.js
@@ -3,12 +3,13 @@ import ActionButton from 'action-button'
 import defined from 'xo-defined'
 import Icon from 'icon'
 import React from 'react'
+import { addSubscriptions, generateRandomId } from 'utils'
 import { alert, confirm } from 'modal'
 import { createRemote, editRemote, subscribeRemotes } from 'xo'
 import { error } from 'notification'
 import { format, parse } from 'xo-remote-parser'
-import { addSubscriptions, linkState, generateRandomId } from 'utils'
 import { injectState, provideState } from '@julien-f/freactal'
+import { linkState } from 'freactal-utils'
 import { map, some, trimStart, keyBy } from 'lodash'
 import { Number } from 'form'
